+++ g4f/Provider/har/__init__.py
@@ -15,7 +15,7 @@ from ..helper import get_last_user_message
 from ..openai.har_file import get_headers
 
 class HarProvider(AsyncGeneratorProvider, ProviderModelMixin):
-    label = "Legacy LM Arena (Har)"
+    label = "LMArena (Har)"
     url = "https://legacy.lmarena.ai"
     api_endpoint = "/queue/join?"
     working = True
