+++ g4f/Provider/not_working/__init__.py
@@ -25,7 +25,6 @@ from .MagickPen         import MagickPen
 from .Phind             import Phind
 from .Poe               import Poe
 from .Prodia            import Prodia
-from .PuterJS           import PuterJS
 from .Raycast           import Raycast
 from .RubiksAI          import RubiksAI
 from .Theb              import Theb
