+++ g4f/Provider/template/__init__.py
@@ -1,3 +1,2 @@
 from .BackendApi import BackendApi
 from .OpenaiTemplate import OpenaiTemplate
-from .Puter import Puter
